[docker] wait-images separate from release-images
