+++ docker/__tests__/release-images.test.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env -S pnpm test release-images.test.js
 
-import { assertTagMatchesSourceVersion, getImageReleaseGroupByImageTagPrefix, isReleaseImage } from '../release-images.mjs';
-
+import { getImageReleaseGroupByImageTagPrefix, } from '../release-images.mjs';
+import { lazyImports, isReleaseImage, assertTagMatchesSourceVersion } from '../image-helpers.js';
 describe('releaseImages', () => {
     it('gets aptos-node as the default image group', () => {
         const prefix = 'image-banana';
@@ -22,7 +22,8 @@ describe('releaseImages', () => {
         expect(isReleaseImage("nightly-banana")).toEqual(false);
         expect(isReleaseImage("aptos-node-v1.2.3")).toEqual(true);
     });
-    it('asserts version match', () => {
+    it('asserts version match', async () => {
+        await lazyImports();
         // toThrow apparently matches a prefix, so this works but it does actually test against the real config version
         // Which... hilariously means this would fail if the version was ever 0.0.0
         expect(() => assertTagMatchesSourceVersion("aptos-node-v0.0.0")).toThrow("image tag does not match cargo version: aptos-node-v0.0.0");
